fix: rename casegeneration form -> case creation form
