+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/components/CaseCreationForm/utils/transformErrors.ts
