--- apps/backoffice-v2/src/pages/Entities/components/CaseCreation/components/CaseGenerationForm/CaseGenerationForm.tsx
@@ -1,35 +1,0 @@
-import { TWorkflowDefinition } from '@/domains/workflows/fetchers';
-import { SubmitSection } from '@/pages/Entities/components/CaseCreation/components/CaseGenerationForm/components/SubmitSection';
-import { useCaseGenerationForm } from '@/pages/Entities/components/CaseCreation/components/CaseGenerationForm/hooks/useCaseGenerationForm';
-import { useFormSchema } from '@/pages/Entities/components/CaseCreation/components/CaseGenerationForm/hooks/useFormSchema';
-import { transformErrors } from '@/pages/Entities/components/CaseCreation/components/CaseGenerationForm/utils/transformErrors';
-import { DynamicForm, baseLayouts } from '@ballerine/ui';
-import { FunctionComponent } from 'react';
-
-interface FormProps {
-  workflowDefinition: TWorkflowDefinition;
-}
-
-const layouts = {
-  ...baseLayouts,
-  ButtonTemplates: {
-    ...baseLayouts.ButtonTemplates,
-    SubmitButton: SubmitSection,
-  },
-};
-
-export const CaseGenerationForm: FunctionComponent<FormProps> = ({ workflowDefinition }) => {
-  const { jsonSchema, uiSchema } = useFormSchema(workflowDefinition);
-  const { isLoading, handleSubmit } = useCaseGenerationForm(workflowDefinition);
-
-  return (
-    <DynamicForm
-      schema={jsonSchema}
-      uiSchema={uiSchema}
-      onSubmit={handleSubmit}
-      disabled={isLoading}
-      transformErrors={transformErrors}
-      layouts={layouts as typeof baseLayouts}
-    />
-  );
-};
